Struct libits_client::reception::exchange::map_extended_message::Lane

pub struct Lane {

    pub id: u64,
    pub signal_id: u64,
    pub approach_id: Option<u32>,
    pub left: bool,
    pub straight: bool,
    pub right: bool,
    pub speed_limit: u16,
    pub ingress: bool,
    pub egress: bool,
    pub geom: Vec<[f32; 2]>,
    pub is_pedestrian_lane: Option<bool>,
    pub is_vehicle_lane: Option<bool>,
    pub is_bus_lane: Option<bool>,
    pub is_bike_lane: Option<bool>,
    pub connections: Vec<Connection>,

}

Fields
id: u64signal_id: u64

ID of the signal, corresponding to the SPAT signal ID

approach_id: Option<u32>

ID of the approach (group of lanes)

left: bool

True if the lane allows a left turn at its end

straight: bool

True if the lane allows to go straight ahead at its end

Does not seem to be present in some MAPEM despite the documentation telling it is mandatory; Defaulting value (as false)

FIXME defaulting to false if missing does not fit reality, replace it by an Option

right: bool

True if this lane allows a right turn at its end

speed_limit: u16

Speed limit on this lane (km/h)

ingress: boolegress: boolgeom: Vec<[f32; 2]>

List of points representing the lane

Each point is an array of numbers where the first value is the longitude and the second value is the latitude

Note: latitude and longitude refers to the WGS84

is_pedestrian_lane: Option<bool>is_vehicle_lane: Option<bool>is_bus_lane: Option<bool>is_bike_lane: Option<bool>connections: Vec<Connection>